+++ lib/uuid_util.dart
@@ -1,12 +1,13 @@
 library uuid_util;
 
 import 'dart:math';
+import 'dart:typed_data';
 
 class UuidUtil {
   /// Math.Random()-based RNG. All platforms, fast, not cryptographically strong. Optional Seed passable.
-  static List<int> mathRNG({int seed = -1}) {
+  static Uint8List mathRNG({int seed = -1}) {
     int rand;
-    final List<int> b = List<int>(16);
+    final Uint8List b = Uint8List(16);
 
     final Random _rand = (seed == -1) ? Random() : Random(seed);
     for (int i = 0; i < 16; i++) {
@@ -20,8 +21,8 @@ class UuidUtil {
   }
 
   /// Crypto-Strong RNG. All platforms, unknown speed, cryptographically strong (theoretically)
-  static List<int> cryptoRNG() {
-    final List<int> b = List<int>(16);
+  static Uint8List cryptoRNG() {
+    final Uint8List b = Uint8List(16);
     final Random rand = Random.secure();
     for (int i = 0; i < 16; i++) {
       b[i] = rand.nextInt(1 << 8);